backend/requirements.txt
Flask
Flask-Cors
pytesseract
Pillow
opencv-python-headless
numpy